renamed packages and added to .gitignore
