+++ .gitignore
@@ -113,3 +113,4 @@ Session.vim
 
 .classpath
 .project
+bin/
