--- src/Interface/Executive/Legacy/Status.java
@@ -1,87 +1,0 @@
-package Interface.Executive.Legacy;
-
-import DataRecorderAndRetriever.*;
-import Interface.*;
-import java.util.*;
-import javafx.geometry.*;
-import javafx.scene.*;
-import javafx.scene.control.*;
-import javafx.scene.input.*;
-import javafx.scene.layout.*;
-import javafx.stage.*;
-/**
- * 
- * @author James Sonne & Devin Matte
- * @version v0.5-Alpha
- * @since 2016-02-11
- */
-public class Status {
-	/**
-	 * Show Item Status Window
-	 * 
-	 * @param file
-	 *            Links the main item File
-	 */
-	public static void show(String file) {
-		ItemManager manage = new ItemManager();
-		Label errors = new Label();
-		Stage window = new Stage();
-		VBox showLayout = new VBox(10), show2Layout = new VBox(10);
-		Scene show = new Scene(showLayout, 250, 250), show2 = new Scene(show2Layout, 500, 150);
-		show2.getStylesheets().add("style.css");
-		Label label = new Label();
-
-		UIManager.windowBasic(window, "Equipment Status", 250, label, errors, show);
-
-		// Show Item Status
-		Label showStatus = new Label();
-		TextField showitemID = new TextField();
-		showitemID.setPromptText("Item ID");
-		Button ShowSubmit = new Button("Submit");
-		ShowSubmit.setOnAction(e -> {
-			errors.setText("");
-			if (UIManager.isInt(showitemID, showitemID.getText())) {
-				manage.getItem(new Scanner(showitemID.getText()).nextInt(), file);
-				showStatus.setText(manage.getItem(new Scanner(showitemID.getText()).nextInt(), file).toString());
-				showitemID.clear();
-				window.setScene(show2);
-			} else {
-				errors.setText("Something Went Wrong. Try Again");
-			}
-		});
-		showitemID.setOnKeyPressed(e -> {
-			errors.setText("");
-			if (e.getCode().equals(KeyCode.ENTER)) {
-				if (UIManager.isInt(showitemID, showitemID.getText())) {
-					manage.getItem(new Scanner(showitemID.getText()).nextInt(), file);
-					showStatus.setText(manage.getItem(new Scanner(showitemID.getText()).nextInt(), file).toString());
-					showitemID.clear();
-					window.setScene(show2);
-				} else {
-					errors.setText("Something Went Wrong. Try Again");
-				}
-			}
-		});
-
-		Button Back = new Button("Back");
-		Back.setOnAction(e -> {
-			errors.setText("");
-			window.close();
-		});
-		Button twoBack = new Button("Back");
-		twoBack.setOnAction(e -> {
-			errors.setText("");
-			window.setScene(show);
-		});
-
-		showLayout.getChildren().addAll(label, errors, showitemID, ShowSubmit, Back);
-		showLayout.setAlignment(Pos.CENTER);
-
-		show2Layout.getChildren().addAll(showStatus, twoBack);
-		show2Layout.setAlignment(Pos.CENTER);
-
-		// Display window and wait for it to be closed before returning
-		window.setScene(show);
-		window.showAndWait();
-	}
-}
