+++ src/us/nhstech/inventory/utils/Item.java
@@ -1,4 +1,4 @@
-package DataRecorderAndRetriever;
+package us.nhstech.inventory.utils;
 
 /**
  *
