+++ src/us/nhstech/inventory/utils/LogManager.java
@@ -1,4 +1,4 @@
-package DataRecorderAndRetriever;
+package us.nhstech.inventory.utils;
 
 import java.util.ArrayList;
